backend/user_profiles/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers

from user_profiles.models import UserProfile


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ["id", "pk", "username", "is_staff", "is_superuser"]


class UserProfileSerializer(serializers.ModelSerializer):
    user = UserSerializer(read_only=True)

    class Meta:
        model = UserProfile
        fields = "__all__"
